change update logic
