+++ src/components/main-controller.js
@@ -1,14 +1,11 @@
-/* eslint-disable no-unused-vars */
 
 import Sort from './sort.js';
 import TripController from './trip-controller';
 import TripInfo from './trip-info.js';
 import Statistics from './statistics.js';
 import Filter from './filter.js';
-import {Position, MILISECONDS_PER_HOUR} from './constants.js';
+import {Position} from './constants.js';
 import {render, unrender, createElement} from './utils.js';
-import Chart from 'chart.js';
-// import ChartDataLabels from 'chartjs-plugin-datalabels';
 import API from './api.js';
 const END_POINT = `https://htmlacademy-es-9.appspot.com/big-trip/`;
 const AUTHORIZATION = `Basic kjhfdKJLfdsf${Math.random()}`;
@@ -17,28 +14,23 @@ class MainController {
   constructor(events, destinations, offers) {
     this.api = new API({endPoint: END_POINT, authorization: AUTHORIZATION});
     this._events = events;
-    this._sort = new Sort();
-    this._filter = new Filter();
-    this._tripInfo = new TripInfo(events);
-    this._statistics = new Statistics(events);
+    this._destinations = destinations;
+    this._offers = offers;
+    this._sortComponent = new Sort();
+    this._filterComponent = new Filter();
+    this._tripInfo = new TripInfo(this._events);
+    this._statisticsComponent = new Statistics(this._events);
     this._onDataChangeInTripController = this._onDataChangeInTripController.bind(this);
     this._tripController = new TripController(this._renderTripDaysContainer(), events, destinations, offers, this._onDataChangeInTripController);
   }
 
   // начальная инициализация
   init() {
-    this._renderStat();
-    this._statistics.hide();
+    this._statisticsComponent.render(this._events);
     this._renderTripInfo();
     this._renderFilter();
     this._tripController.init();
-  }
-
-  loadEvents() {
-    return this.api.getEvents()
-      .then((events) => {
-        this._events = events;
-      });
+    this._initTabSwitch();
   }
 
   // подготовка контейнера для событий
@@ -52,7 +44,7 @@ class MainController {
   // рендер фильтров FUTURE / PAST / EVERYTHING и инициализация обработчиков нажатия на кнопки
   _renderFilter() {
     const filterContainer = document.querySelector(`.trip-main__trip-controls`);
-    const element = this._filter.getElement();
+    const element = this._filterComponent.getElement();
     render(filterContainer, element, Position.BEFOREEND);
     const filterInputs = [...element.querySelectorAll(`.trip-filters__filter-input`)];
 
@@ -74,191 +66,6 @@ class MainController {
     };
 
     filterInputs.forEach((input) => input.addEventListener(`click`, onFilterTabClick));
-  }
-
-  // рендер статистики
-  _renderStat() {
-    if (this._statistics._element) {
-      unrender(this._statistics);
-    }
-    const element = this._statistics.getElement();
-    const mainPageContainer = document.querySelector(`main .page-body__container`);
-    render(mainPageContainer, element, Position.BEFOREEND);
-    const moneyCtx = element.querySelector(`.statistics__chart--money`);
-    const transportCtx = element.querySelector(`.statistics__chart--transport`);
-    const timeSpendCtx = document.querySelector(`.statistics__chart--time`);
-
-    const moneyStat = {
-      fly: 0,
-      stay: 0,
-      drive: 0,
-      look: 0,
-      eat: 0,
-      ride: 0
-    };
-
-    const transportStat = {
-      drive: 0,
-      ride: 0,
-      fly: 0,
-      sail: 0,
-    };
-
-    const timeStat = {
-    };
-
-    this._events.forEach((event) => {
-      timeStat[event.type] = timeStat[event.type] === undefined ?
-        Math.round((event.endDate - event.startDate) / MILISECONDS_PER_HOUR) : timeStat[event.type] + Math.round((event.endDate - event.startDate) / MILISECONDS_PER_HOUR);
-      if (event.type === `flight`) {
-        moneyStat.fly += event.cost;
-        transportStat.fly += 1;
-      }
-      if (event.type === `check-in`) {
-        moneyStat.stay += event.cost;
-      }
-      if (event.type === `drive`) {
-        moneyStat.drive += event.cost;
-        transportStat.drive += 1;
-      }
-      if (event.type === `sightseeing`) {
-        moneyStat.look += event.cost;
-      }
-      if (event.type === `restaurant`) {
-        moneyStat.eat += event.cost;
-      }
-      if (event.type === `bus` || event.type === `taxi`) {
-        moneyStat.ride += event.cost;
-        transportStat.ride += 1;
-      }
-      if (event.type === `ship`) {
-        transportStat.sail += 1;
-      }
-    });
-
-    const moneyChart = new Chart(moneyCtx, {
-      type: `horizontalBar`,
-      data: {
-        labels: [...Object.keys(moneyStat)],
-        datasets: [{
-          label: `MONEY`,
-          data: [...Object.values(moneyStat)],
-          backgroundColor: `white`,
-          borderWidth: 0,
-        }]
-      },
-      options: {
-        scales: {
-          xAxes: [{
-            barThickness: 5,
-            gridLines: {
-              display: false
-            },
-            ticks: {
-              enabled: false,
-            }
-          }],
-          yAxes: [{
-            gridLines: {
-              display: false
-            },
-            ticks: {
-              enabled: true,
-              fontSize: 18,
-            }
-          }]
-        },
-      },
-    });
-
-    const transportChart = new Chart(transportCtx, {
-      type: `horizontalBar`,
-      data: {
-        labels: [...Object.keys(transportStat)],
-        datasets: [{
-          label: `TRANSPORT`,
-          data: [...Object.values(transportStat)],
-          backgroundColor: `white`,
-          borderWidth: 0,
-        }]
-      },
-      options: {
-        scales: {
-          xAxes: [{
-            barThickness: 5,
-            gridLines: {
-              display: false
-            },
-          }],
-          yAxes: [{
-            gridLines: {
-              display: false
-            },
-            ticks: {
-              enabled: true,
-              fontSize: 18,
-            }
-          }]
-        }
-      },
-    });
-
-    const timeChart = new Chart(timeSpendCtx, {
-      type: `horizontalBar`,
-      data: {
-        labels: [...Object.keys(timeStat)],
-        datasets: [{
-          label: `TIME`,
-          data: [...Object.values(timeStat)],
-          backgroundColor: `white`,
-          borderWidth: 0,
-        }]
-      },
-      options: {
-        scales: {
-          xAxes: [{
-            barThickness: 5,
-            gridLines: {
-              display: false
-            },
-          }],
-          yAxes: [{
-            gridLines: {
-              display: false
-            },
-            ticks: {
-              enabled: true,
-              fontSize: 18,
-            }
-          }]
-        }
-      },
-    });
-
-    const onMenuTabClick = (evt) => {
-      const tabName = evt.target.innerText;
-      switch (tabName) {
-        case `Table`:
-          this._tripController.show();
-          this._statistics.hide();
-          menuBtnTable.classList.add(`trip-tabs__btn--active`);
-          menuBtnStats.classList.remove(`trip-tabs__btn--active`);
-          break;
-        case `Stats`:
-          this._tripController.hide();
-          this._statistics.show();
-          menuBtnStats.classList.add(`trip-tabs__btn--active`);
-          menuBtnTable.classList.remove(`trip-tabs__btn--active`);
-          break;
-        default:
-          break;
-      }
-    };
-
-    const menuBtnTable = document.querySelector(`.trip-tabs__btn--Table`);
-    const menuBtnStats = document.querySelector(`.trip-tabs__btn--Stats`);
-    const menuBtns = [menuBtnTable, menuBtnStats];
-    menuBtns.forEach((btn) => btn.addEventListener(`click`, onMenuTabClick));
   }
 
   // рендер информации о поездке
@@ -294,16 +101,44 @@ class MainController {
       default:
         break;
     }
-    this._renderTripInfo();
-    this._renderStat();
   }
 
   update() {
     this.api.getEvents()
       .then((events) => {
-        this._tripController._events = events;
+        this._events = events;
         this._tripController.renderDays(true);
+        this._statisticsComponent.render();
+        this._renderTripInfo();
       });
+  }
+
+  _initTabSwitch() {
+    const onMenuTabClick = (evt) => {
+      const tabName = evt.target.innerText;
+      switch (tabName) {
+        case `Table`:
+          this._tripController.show();
+          this._statisticsComponent.hide();
+          menuBtnTable.classList.add(`trip-tabs__btn--active`);
+          menuBtnStats.classList.remove(`trip-tabs__btn--active`);
+          break;
+        case `Stats`:
+          this._tripController.hide();
+          this._statisticsComponent.render(this._events);
+          this._statisticsComponent.show();
+          menuBtnStats.classList.add(`trip-tabs__btn--active`);
+          menuBtnTable.classList.remove(`trip-tabs__btn--active`);
+          break;
+        default:
+          break;
+      }
+    };
+
+    const menuBtnTable = document.querySelector(`.trip-tabs__btn--Table`);
+    const menuBtnStats = document.querySelector(`.trip-tabs__btn--Stats`);
+    const menuBtns = [menuBtnTable, menuBtnStats];
+    menuBtns.forEach((btn) => btn.addEventListener(`click`, onMenuTabClick));
   }
 }
